:truck: moved domain specific person classes into package
