+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/MicroServiceApplication.java
@@ -3,11 +3,13 @@ package de.muenchen.raumreservierung;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
  * Application class for starting the microservice.
  */
 @SpringBootApplication
+@EnableScheduling
 @ConfigurationPropertiesScan
 @SuppressWarnings("PMD.UseUtilityClass")
 public class MicroServiceApplication {
