+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/ExternalPersonRepository.java
@@ -1,6 +1,8 @@
 package de.muenchen.raumreservierung.person;
 
 import java.util.UUID;
+
+import de.muenchen.raumreservierung.person.domain.ExternalPerson;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
