+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/InternalPersonRepository.java
@@ -1,6 +1,8 @@
 package de.muenchen.raumreservierung.person;
 
 import java.util.UUID;
+
+import de.muenchen.raumreservierung.person.domain.InternalPerson;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
