+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/domain/ExternalPerson.java
@@ -1,8 +1,10 @@
-package de.muenchen.raumreservierung.person;
+package de.muenchen.raumreservierung.person.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+
 import java.io.Serial;
+
 import lombok.Getter;
 import lombok.Setter;
