+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/domain/PersonType.java
@@ -1,4 +1,4 @@
-package de.muenchen.raumreservierung.person;
+package de.muenchen.raumreservierung.person.domain;
 
 public enum PersonType {
     INTERNAL,
