+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/dto/ExternalPersonRequestDto.java
@@ -1,7 +1,7 @@
 package de.muenchen.raumreservierung.person.dto;
 
-import de.muenchen.raumreservierung.person.PersonTitle;
-import de.muenchen.raumreservierung.person.PersonType;
+import de.muenchen.raumreservierung.person.domain.PersonTitle;
+import de.muenchen.raumreservierung.person.domain.PersonType;
 import jakarta.validation.constraints.NotNull;
 
 public record ExternalPersonRequestDto(
