+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/person/dto/InternalPersonResponseDto.java
@@ -1,8 +1,9 @@
 package de.muenchen.raumreservierung.person.dto;
 
-import de.muenchen.raumreservierung.person.PersonTitle;
-import de.muenchen.raumreservierung.person.PersonType;
+import de.muenchen.raumreservierung.person.domain.PersonTitle;
+import de.muenchen.raumreservierung.person.domain.PersonType;
 import jakarta.validation.constraints.NotNull;
+
 import java.util.UUID;
 
 public record InternalPersonResponseDto(
